🐛 fix(Product): add force-dynamic to ProductList
